Area Selection: Improvements to creating the first box
When creating the first box of a selection, the origin is now selected by
default, and the first corner is offset by one block to make it
immediately selectable too, without first moving the origin out of the
way. Also the box is now by default named based on the area's name.