Store the command-based paste task state. Always set all blocks in projects.
The paste task now also sets the sendCommandFeedback gamerule to false before
the task starts executing and sets it back to true after.
That should be improved to check the initial value, but it requires
waiting for a chat response from the server before the task can start.
So for now it's just always first set to false and then restored to true.