Pick Block improvements/changes/clean-up
The main change is to not pick block anything in the pickBlockLast mode,
if the last traced schematic position is "floating", ie. you wouldn't
place a block into that position when doing the use action.
This prevents changing the held item if you try to place something through
the schematic, but not into an actual schematic position.