Rewrite the block state palette/container resizing
This should fix the old bug of the container possibly getting messed up when
reloading from file after a Rebuild replace operation, if the number of blocks
in the palette crossed a 2^n boundary.

This is because crossing the boundary would cause the palette and backing bit
array to be resized to bitSize + 1. But if some of the old states are then
later completely removed from the container, causing the number of unique
states to fall back below the old 2^n limit, then the stored palette would be
smaller than the corresponding bit array entry size should be.
This is because all unused states would vanish from the new palette when the
backing array was copied over to the new array, requesting IDs from the
new palette as needed. And the bit size of the palette/backing array entry is
not saved explicitly, it is calculated from the size of the stored palette.
So if the palette size fell back below the once-crossed 2^n boundary,
then the container would be read from file with the wrong bit array entry size
the next time, as the size is calculated from the stored palette size.