Restore always assigning ID 0 for air
This fixes multi-region -> single region schematic conversion,
by removing the need to pre-fill the single-region schematic with air.
Without this change it was likely that another block would get ID 0,
which would cause the single-region schematic to be filled with that
block instead of air in places where the multi-region schematic didn't
have any regions.