Move the attack and use hooks to the click handling code. Add pickBlockAuto.
This fixes use actions that start from the outside of schematic regions
getting through and not being handled by the Easy Place or Placement
Restriction code.

There is now a new pickBlockAuto option, which does the equivalent of the old
pickBlockLast from the click handling code instead, which means that it also
happens without releasing the use button and clicking again, so in other
words this new variant works when dragging the cursor over while holding down
the use key.