Finish the update to 20w45a
There are some changes that are kind of half assed for now, which should
be done better later on, probably by rewriting a bunch of the schematic
world and chunk management stuff starting from 1.12.2 to try to unify
it better between versions, as well as simplify it for the type of use
that mostly happens in this mod vs. how vanilla works.