Rewrite parts of the renderer and placement manager...
to handle rendering in full chunk high sections instead of
16x16x16 sections ("sub-chunks"). This further reduces the amount
of buffer allocated on the GPU.

Note: Fluid rendering is broken due to the Mojank fluid renderer
masking the position with `& 0xF`.

Either the fluid renderer needs to be copy pasted, or a hack fix
for rendering fluids in 16-high sections needs to be used.
In the latter case might as well just render everything in those
smaller passes instead of separating fluids.